added tests for habit and completion API
