+++ src/routes/api/user.ts
@@ -2,6 +2,7 @@ import Koa from "koa";
 import Router from "koa-router";
 import { isAuthenticated } from "../../auth";
 import User from "../../models/User";
+import Completion from "../../models/Completion";
 
 export default (router: Router) => {
   // Sanity check function
@@ -11,5 +12,23 @@ export default (router: Router) => {
       .populate("habits");
 
     if (user) ctx.body = { response: user };
+  });
+
+  router.get("/user/weekly", isAuthenticated(), async (ctx: Koa.Context) => {
+    const user = await User.findById(ctx.state.user)
+      .select("habits")
+      .populate("habits");
+
+    const curr = new Date();
+    const first = curr.getDate() - curr.getDay(); // First day is the day of the month - the day of the week
+    const firstDayOfTheWeek = new Date(curr.setDate(first + 1)); // add 1 to start on Monday
+
+    const completions = await Completion.find({
+      user: ctx.state.user,
+      date: { $gte: firstDayOfTheWeek },
+    }).select("habits date");
+
+    ctx.status = 200;
+    ctx.body = { habits: user.habits, completions };
   });
 };
